+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 0.15.0 (in development)
 -----------------------
 
+* Add new optional arguments ``--genes`` and ``--exclude`` to :command:`prepare-depth-of-coverage` command.
+
 0.14.0 (2022-04-03)
 -------------------